+++ src/main/java/com/app/flashgram/post/appication/interfaces/CommentCommandRepository.java
@@ -1,0 +1,9 @@
+package com.app.flashgram.post.appication.interfaces;
+
+import com.app.flashgram.post.domain.comment.Comment;
+
+public interface CommentCommandRepository {
+
+    Comment save(Comment comment);
+    Comment findById(Long id);
+}
